use case insensitive emails
Warning: There is no guarantee for emails to be case insensitive,
but case sensitivity on emails is particularly annoying and
email servers that implement case sensitivity are extremely rare.

This patch proposes to store the original email given by the user,
but not to distinguish `A@x.x` from `a@x.x` for any query in the
database.

modified:   src/eba_db.ml
modified:   template.distillery/PROJECT_NAME.sql